app/requirements.txt
Flask==3.0.3
discord.py==2.4.0
python-dotenv